+++ packages/api/src/AccessRegistryApi/index.ts
@@ -1,0 +1,2 @@
+export * from './AccessRegistryApi';
+export type { AccessRegistryEntity } from './types';