+++ packages/ui/src/components/Layout/Layout.tsx
@@ -21,7 +21,13 @@ export const Layout = ({
   headerRight,
   header = <LayoutHeader title={title} fullPage={fullPage} rightElement={headerRight} />,
 }: LayoutProps) => (
-  <Container disableGutters={fullPage} maxWidth={!fullPage && 'xl'} component={Stack} spacing={disablePaddings ? 0 : 4}>
+  <Container
+    disableGutters={fullPage}
+    minWidth={800}
+    maxWidth={!fullPage && 'xl'}
+    component={Stack}
+    spacing={disablePaddings ? 0 : 4}
+  >
     {header}
     <Content>{children}</Content>
   </Container>